feat: Beta Version 4
Pretty much complete rework of the codebase, still functions in a
relatively similar way, just made the parsing make more sense and be in
less convoluted ways.

NOTE: Formatting is no longer an extra step, it's a part of the parsing as it
should be.

NOTE: Reader now uses a more simple and deterministic way of restoring
context for errors by showing a fixed number of lines before the
highlighted area.

NOTE: Tests now allow for GO_DEBUG=1 to print the JSON of output tree.
Really helpful to understanding what the parser has actually generated.